+++ src/FusionRpg.Core/StatMath.cs
@@ -4,16 +4,16 @@ namespace FusionRpg.Core;
 
 public static class StatMath
 {
-    public static int ScaleHpOrAtk(int baseline, float percent, int flat) =>
-        Math.Max(1, (int)Math.Round(baseline * percent) + flat);
+    public static long ScaleHpOrAtk(long baseline, float percent, long flat) =>
+        Math.Max(1L, (long)Math.Round(baseline * percent) + flat);
 
     public static long ScaleHp(long baseline, float percent, int flat) =>
         Math.Max(1L, (long)Math.Round(baseline * percent) + flat);
 
-    public static int ScaleIncoming(int damage, float defensePercent, int defenseFlat)
+    public static long ScaleIncoming(long damage, float defensePercent, long defenseFlat)
     {
         var p = defensePercent <= 0 ? 1f : defensePercent;
-        return Math.Max(0, (int)Math.Round(damage / (double)p) - defenseFlat);
+        return Math.Max(0L, (long)Math.Round(damage / (double)p) - defenseFlat);
     }
 
     public static bool IsIdentity(StatMod m) =>
